+++ src/arrays/noHitBosses.ts
@@ -1,8 +1,10 @@
 import { BossID } from "isaac-typescript-definitions";
-import { ReadonlySet } from "isaacscript-common";
+import { ReadonlyMap, ReadonlySet } from "isaacscript-common";
 import { BOSS_IDS } from "../cachedEnums";
 import { IS_DEV } from "../constants";
 import { BossIDCustom } from "../enums/BossIDCustom";
+
+const DEFAULT_NO_HIT_BOSS_MINUTES = 2;
 
 /** @see about.md */
 const NO_HIT_EXCEPTION_BOSSES = new ReadonlySet([
@@ -19,19 +21,64 @@ export const NO_HIT_BOSSES: readonly BossID[] = BOSS_IDS.filter(
   // eslint-disable-next-line unicorn/prefer-spread
 ).concat(...Object.values(BossIDCustom));
 
-const HARD_BOSS_IDS = new ReadonlySet<BossID>([
+const BOSSES_WITH_CUSTOM_MINUTES = new ReadonlyMap<BossID, float>([
+  // Too easy.
+  [BossID.MONSTRO, 1], // 1
+
+  // Too easy.
+  [BossID.LARRY_JR, 1], // 2
+
+  // Mostly easy.
+  [BossID.FAMINE, 1.5], // 9
+
+  // Too easy.
+  [BossID.DUKE_OF_FLIES, 1], // 13
+
+  // Too easy.
+  [BossID.GEMINI, 1], // 17
+
+  // Mostly easy.
+  [BossID.FISTULA, 1.5], // 18
+
+  // Too easy.
+  [BossID.STEVEN, 1], // 20
+
+  // Mostly easy.
+  [BossID.TERATOMA, 1.5], // 33
+
+  // Mostly easy.
+  [BossID.PIN, 1.5], // 37
+
+  // Mostly easy.
+  [BossID.DINGLE, 1.5], // 44
+
+  // Mostly easy.
+  [BossID.GURGLING, 1.5], // 56
+
   // When play testing, even without dealing any damage to Brownie, he kills himself at around the 1
   // minute and 10 seconds mark.
-  BossID.BROWNIE, // 58
+  [BossID.BROWNIE, 1], // 58
+
+  // Mostly easy.
+  [BossID.LITTLE_HORN, 1.5], // 60
+
+  // Mostly easy.
+  [BossID.RAG_MAN, 1.5], // 61
+
+  // Mostly easy.
+  [BossID.DANGLE, 1.5], // 64
+
+  // Mostly easy.
+  [BossID.TURDLING, 1.5], // 65
 
   // The boss is extremely difficult to do without taking any damage.
-  BossID.HORNFEL, // 82
+  [BossID.HORNFEL, 1], // 82
 
   // The boss is extremely difficult to do without taking any damage.
-  BossID.SCOURGE, // 85
+  [BossID.SCOURGE, 1], // 85
 
   // 2 minutes is probably doable but it is pretty difficult.
-  BossID.ROTGUT, // 87
+  [BossID.ROTGUT, 1], // 87
 ]);
 
 export function getNumMinutesForBossObjective(bossID: BossID): int {
@@ -39,5 +86,6 @@ export function getNumMinutesForBossObjective(bossID: BossID): int {
     return 0.05;
   }
 
-  return HARD_BOSS_IDS.has(bossID) ? 1 : 2;
+  const customMinutes = BOSSES_WITH_CUSTOM_MINUTES.get(bossID);
+  return customMinutes ?? DEFAULT_NO_HIT_BOSS_MINUTES;
 }